requirements.txt
Faker>=20.1.0
pandas>=1.3.1
SQLAlchemy<2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
mysqlclient>=2.2.0